Finish logged-in user's preferences to update the query and non-logged users to query everything
